Move transaction to its own module
